Module debris_mir::mir_builder

Structs

MirBuilder

MirSingletons

Holds some singletons objects for easier access

Enums

ReturnContextBehavior 🔒

Simple enum to specify how to calculate the return context

Functions

exists_runtime_context 🔒

Returns true if any context, that lies ‘between’ target_id and current_context, is marked as runtime context. This is used to check whether a variable update can be performed at compile time

get_context 🔒

get_context_mut 🔒

get_entry_context 🔒