Module debris_mir::mir_nodes

Macros

mir_node_declaration

Structs

Branch

An if-statement which has a condition, a return value and two possible branches

FunctionCall

Goto

Goes to a specific context

PrimitiveDeclaration

Declares a primitive value and assigns it to target

PropertyAccess

Binds a required property of an object to an object id so it can be used

RuntimeCopy

Tries to copy a value (Does nothing if the value is comptime)

RuntimePromotion

Tries to promote a comptime value to a runtime value, but does not fail if no such conversion exists. Used in let foo = expression(); statements, because foo by default stores runtime encodable values.

VariableUpdate

Updates the variable at target to value

VerifyPropertyExists

VerifyTupleLength

VerifyValueComptime

Enums

MirNode